fix(object-runtime): settle string helper shifts before object runtime
Move native-string finalize shift reconciliation to the native string module so the standalone object runtime can settle any pending import-base drift before registering its helper bodies.

This keeps __obj_find's baked __obj_hash call aligned with the i32-returning helper instead of being over-shifted onto an externref helper when dynamic property programs combine native strings with object runtime emission.
