+++ src/codegen/expressions/late-imports.ts
@@ -14,6 +14,8 @@ import { addFuncType } from "../registry/types.js";
 import { addUnionImportsViaRegistry } from "../shared.js";
 import { ensureObjectRuntime, OBJECT_RUNTIME_HELPER_NAMES } from "../object-runtime.js";
 
+export { reconcileNativeStrFinalizeShift } from "../native-strings.js";
+
 /**
  * #1471: helper names that `addUnionImports` provides Wasm-native
  * implementations for under no-JS-host mode (WASI / standalone). When any of
@@ -327,123 +329,6 @@ export function ensureLateImport(
   const typeIdx = addFuncType(ctx, paramTypes, resultTypes);
   addImport(ctx, "env", name, { kind: "func", typeIdx });
   return ctx.funcMap.get(name);
-}
-
-/**
- * #1677: reconcile native-string helper function indices at the end of the
- * import-collection finalize phase.
- *
- * The native-string runtime helpers (`__str_flatten` & co.) are emitted as
- * DEFINED functions *mid-finalize* (the first `ensureNativeStringHelpers`
- * call), recording `ctx.nativeStrHelperImportBase` = the import count at that
- * instant. Finalize then continues and adds more imports via raw `addImport`
- * (string methods, parseInt, Promise statics, iterator/generator bridges, …).
- * Raw `addImport` does NOT shift defined-function indices (the #618 revert
- * deliberately removed that), so every helper body's internal `call` to a
- * sibling helper — and every `nativeStrHelpers` map entry — is now stale-low by
- * exactly `(numImportFuncs - base)`.
- *
- * This reconciliation runs once, after finalize is otherwise complete, and
- * applies that single uniform delta via the same `shiftLateImportIndices`
- * walker the compilation-phase late-import path uses — unifying the two shift
- * regimes (#1666 root cause). It is a no-op unless native-string helpers were
- * emitted, so the default GC path is never touched; and because it shifts
- * `funcIdx >= base` (sibling-helper calls) but not `funcIdx < base` (calls to
- * imports added before the helpers), it cannot corrupt the Math.* host
- * trampoline path that the #618 naive `addImport` shift broke.
- */
-export function reconcileNativeStrFinalizeShift(ctx: CodegenContext): void {
-  const base = ctx.nativeStrHelperImportBase;
-  if (base < 0) return; // helpers never emitted (default GC path or no strings)
-  const added = ctx.numImportFuncs - base;
-  // Re-base for the next incremental call so repeated invocations only apply
-  // the NEW imports added since the previous reconcile. (Imports are inserted
-  // between finalize and body compilation at several points — the
-  // register-prototype block, the deferred WASI helper emitters, etc. — and
-  // each batch drifts the indices further.)
-  ctx.nativeStrHelperImportBase = ctx.numImportFuncs;
-  if (added <= 0) return;
-
-  // Uniform defined-function index shift, restricted to the native-string
-  // finalize regime.
-  //
-  // At the point this runs (mid/end of finalize, before `collectDeclarations`),
-  // EVERY function already in `ctx.mod.functions` is an eagerly-emitted runtime
-  // helper (the native-string helpers AND their dependencies: `__box_number`,
-  // `__unbox_number`, `__toUint32`, …). They were all emitted while
-  // `numImportFuncs == base`, so their absolute index is `base + arrayPos` and
-  // any `call`/`ref.func funcIdx >= base` they baked points at a *sibling*
-  // defined function under that same regime. Adding `added` finalize imports
-  // (string methods, parseInt, Promise statics, …) bumps every defined
-  // function's true absolute index by `added` but leaves the baked call targets
-  // and the `funcMap` entries stale-low. We repair ALL of them uniformly by
-  // `added` for `funcIdx >= base`.
-  //
-  // Earlier versions shifted only `nativeStrHelpers`-named functions; that
-  // missed dependency helpers like `__box_number` (registered via
-  // `addUnionImportsAsNativeFuncs`, not tracked in `nativeStrHelpers`), whose
-  // stale-low `funcMap` entry made later callers (e.g. `__vec_get`) target the
-  // wrong function — `call[k] not enough arguments on the stack`.
-  //
-  // #618 safety: gated on `base >= 0`, which is set only inside
-  // `ensureNativeStringHelpers` (native-strings path: wasi/standalone or
-  // explicit `--nativeStrings`). On the default JS-host GC path the helpers are
-  // never emitted, `base` stays -1, and this is a hard no-op — so the
-  // Math.*-trampoline corruption the #608 `addImport` shift caused cannot recur.
-  // The `funcIdx < base` guard further protects calls into pre-helper imports
-  // (the host trampolines) from ever being shifted.
-  const seen = new Set<Instr[]>();
-  function shiftBody(instrs: Instr[]): void {
-    if (seen.has(instrs)) return;
-    seen.add(instrs);
-    for (const instr of instrs) {
-      const a = instr as any;
-      if (
-        (instr.op === "call" || instr.op === "return_call" || instr.op === "ref.func") &&
-        typeof a.funcIdx === "number" &&
-        a.funcIdx >= base
-      ) {
-        a.funcIdx += added;
-      }
-      if (Array.isArray(a.body)) shiftBody(a.body);
-      if (Array.isArray(a.then)) shiftBody(a.then);
-      if (Array.isArray(a.else)) shiftBody(a.else);
-      if (Array.isArray(a.catches)) {
-        for (const c of a.catches) if (Array.isArray(c.body)) shiftBody(c.body);
-      }
-      if (Array.isArray(a.catchAll)) shiftBody(a.catchAll);
-    }
-  }
-  for (const fn of ctx.mod.functions) {
-    shiftBody(fn.body);
-  }
-  // Shift `funcMap` entries for DEFINED functions only. The `added` imports we
-  // just inserted now occupy indices `[base, base+added)`, which collides by
-  // value with stale defined-function entries (also `>= base`). We cannot
-  // disambiguate by index value, so we gate on the function NAME being a
-  // defined function (present in `ctx.mod.functions`). Import names are never
-  // in that set, so their (correct) entries are left untouched.
-  const definedNames = new Set<string>();
-  for (const fn of ctx.mod.functions) {
-    const n = (fn as { name?: string }).name;
-    if (n) definedNames.add(n);
-  }
-  for (const [name, idx] of ctx.funcMap) {
-    if (idx >= base && definedNames.has(name)) ctx.funcMap.set(name, idx + added);
-  }
-  // Keep the helper-name map (read directly by string lowering) in lockstep —
-  // every entry is a defined function by construction.
-  for (const [name, idx] of ctx.nativeStrHelpers) {
-    if (idx >= base) ctx.nativeStrHelpers.set(name, idx + added);
-  }
-  // Shift export descriptors. Exports only ever reference defined functions in
-  // this regime (helpers/runtime exports like `__vec_get`); a func export with
-  // `index >= base` is a defined function whose true slot moved by `added`.
-  for (const exp of ctx.mod.exports) {
-    if (exp.desc.kind === "func" && exp.desc.index >= base) {
-      exp.desc.index += added;
-    }
-  }
 }
 
 /**
